requirements.txt

Django==4.2.15
djangorestframework==3.15.2
django-cors-headers==4.4.0
pyserial==3.5
